Service_System/djangoenv/requirements.txt
Django==4.2.25
djangorestframework==3.14.0
Pillow==10.0.0

